Support for downloading unreleased dev builds
Core packages, like webdev/dwds, need to run tests again dev candidate releases that
will be rolled into flutter.

This change allows users of the CI to specify dev_unsigned to download `dev` builds
that are not released. One caveat of this approach is that `dev_unsigned` is not a channel
and therefore specifying a specific version can be clunky. For example, one needs to specify
`2.14.0-dev_unsigned.1.2` to download a specific raw dev build.

Two alternatives:
* keep the channels as is, but add an extra argument to specify whether to use a
  raw build.
* always use the raw build (presumably under linux machines only)

Fixes #29